back/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - PostgreSQL
asyncpg==0.29.0
SQLAlchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Database - Elasticsearch
elasticsearch==8.11.0
elasticsearch-async==6.2.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Celery
celery==5.3.6
flower==2.0.1

# Object Storage
minio==7.2.3
boto3==1.34.27

# Email Processing
aiosmtplib==3.0.1
email-validator==2.1.0
imapclient==3.0.1
mail-parser==3.15.0
beautifulsoup4==4.12.3
lxml==5.1.0

# File Parsing
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1
pdfplumber==0.10.3
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0
python-docx==1.1.0

# AI / LLM
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
tiktoken==0.5.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==42.0.0

# Validation
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-slugify==8.0.3
phonenumbers==8.13.29
chardet==5.2.0
python-magic==0.4.27
jinja2==3.1.3

# Monitoring & Logging
sentry-sdk==1.40.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
